@@ -39,6 +39,7 @@ namespace TalkToApi.V1.Controllers
             _tokenRepository = tokenRepository;
         }
 
+        [ApiVersion("1.2")]
         [Authorize]
         [HttpGet("", Name = "UsuarioObterTodos")]
         [DisableCors()]
@@ -251,7 +252,7 @@ namespace TalkToApi.V1.Controllers
                 new Claim(JwtRegisteredClaimNames.Sub, usuario.Id)
             };
 
-            var key = new SymmetricSecurityKey(Encoding.UTF8.GetBytes("chave-api-jwt-minhas-tarefas")); //Recomendo -> appsettings.json
+            var key = new SymmetricSecurityKey(Encoding.UTF8.GetBytes("Cm2FxFNb2Rgx1IrPoI8M6cC1IcutDawX")); //Recomendo -> appsettings.json
             var sign = new SigningCredentials(key, SecurityAlgorithms.HmacSha256);
             var exp = DateTime.UtcNow.AddHours(1);
 
